+++ dflib/src/main/java/org/dflib/agg/Percentiles.java
@@ -191,7 +191,7 @@ public class Percentiles {
 
                 BigDecimal lbd = new BigDecimal(sorted.get(lower));
                 BigDecimal ubd = new BigDecimal(sorted.get(upper));
-                BigDecimal fraction = new BigDecimal(di - lower);
+                BigDecimal fraction = BigDecimal.valueOf(di - lower);
 
                 // sorted[lower] + (di - lower) * (sorted[upper] - sorted[lower])
                 return lbd.add(ubd.subtract(lbd).multiply(fraction));
@@ -223,7 +223,7 @@ public class Percentiles {
 
                 BigDecimal lbd = sorted.get(lower);
                 BigDecimal ubd = sorted.get(upper);
-                BigDecimal fraction = new BigDecimal(di - lower);
+                BigDecimal fraction = BigDecimal.valueOf(di - lower);
 
                 // sorted[lower] + (di - lower) * (sorted[upper] - sorted[lower])
                 return lbd.add(ubd.subtract(lbd).multiply(fraction));